bug fixed: there's totally no problems in boot add: logger(use macro 'NDEBUG' to switch) add: switchable boot loader add: new cpu init method(after boot) deleted: cyclecounter

+++ common.cpp
@@ -3,7 +3,7 @@
 //
 #include "common.h"
 Byte getBit(const Byte b, const Byte p){
-    return (b >> p) & 1;
+    return static_cast<Byte>((b >> p) & 1);
 }
 void setBit(Byte & b, const Byte p){
     b |= (1 << p);
